updated to resolve comments
