+++ jaseci_ai_kit/jaseci_ai_kit/modules/tts/config.cfg
@@ -10,3 +10,4 @@ hifigan = speechbrain/tts-hifigan-libritts-22050Hz
 [HPARAMS]
 rate = 22050
 denoicer_strength = 0.05
+force_reload = 0
